+++ mod-di-converter-storage-server/src/main/java/org/folio/services/ProfileService.java
@@ -108,4 +108,20 @@ public interface ProfileService<T, S, D> {
    * @return - profile name
    */
   String getProfileName(T profile);
+
+  /**
+   * Checks is profile contains child profiles
+   *
+   * @param profile  - T entity
+   * @return - boolean value. True if profile contains child profiles
+   */
+  Future<Boolean> isProfileContainsChildProfiles(T profile);
+
+  /**
+   * Checks is profile contains parent profiles
+   *
+   * @param profile  - T entity
+   * @return - boolean value. True if profile contains parent profiles
+   */
+  Future<Boolean> isProfileContainsParentProfiles(T profile);
 }