+++ mod-di-converter-storage-server/src/main/resources/templates/db_scripts/defaultData/default_update_marc_authority_job_profile.sql
@@ -32,6 +32,7 @@ SET jsonb =  '{
   "action": "CREATE",
   "deleted": false,
   "hidden": false,
+  "remove9Subfields": false,
   "description": "This action profile is used with FOLIO''s default job profile for creating a MARC authority record in source-record-storage. It cannot be edited, duplicated, or deleted.",
   "folioRecord": "AUTHORITY",
   "childProfiles": [],